LayoutTests/crypto/subtle/hmac-generate-export-key-jwk-sha224.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<script src="../resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Test exporting a SHA-224 HMAC key with JWK format");

jsTestIsAsync = true;

var extractable = true;

debug("Generating a key...");
crypto.subtle.generateKey({name: "hmac", hash: "sha-224"}, extractable, ["sign", "verify"]).then(function(key) {
    debug("Exporting a key...");
    return crypto.subtle.exportKey("jwk", key);
}).then(function(result) {
    key = result;

    shouldBe("key.kty", "'oct'");
    shouldBe("key.key_ops", "['sign', 'verify']");
    shouldBe("key.alg", "'HS224'");
    shouldBe("key.ext", "true");
    shouldBe("Base64URL.parse(key.k).byteLength", "64");

    finishJSTest();
});
</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
